requirements.txt
fastapi
psycopg2-binary
pydantic
uvicorn
httpx
asyncpg